+++ replacer-backend/src/main/java/es/bvalero/replacer/common/domain/ReplacementType.java
@@ -17,16 +17,10 @@ public class ReplacementType {
     public static final ReplacementType DATE_UPPERCASE = ReplacementType.of(ReplacementKind.DATE, "Mes en mayúscula");
     public static final ReplacementType DATE_UNORDERED = ReplacementType.of(ReplacementKind.DATE, "Fecha desordenada");
 
-    public static final ReplacementType ACUTE_O_NUMBERS = ReplacementType.of(
-        ReplacementKind.COMPOSED,
-        "ó entre números"
-    );
-    public static final ReplacementType ACUTE_O_WORDS = ReplacementType.of(
-        ReplacementKind.COMPOSED,
-        "ó entre palabras"
-    );
+    public static final ReplacementType ACUTE_O_NUMBERS = ReplacementType.of(ReplacementKind.STYLE, "ó entre números");
+    public static final ReplacementType ACUTE_O_WORDS = ReplacementType.of(ReplacementKind.STYLE, "ó entre palabras");
 
-    public static final ReplacementType CENTURY = ReplacementType.of(ReplacementKind.COMPOSED, "Siglo sin versalitas");
+    public static final ReplacementType CENTURY = ReplacementType.of(ReplacementKind.STYLE, "Siglo sin versalitas");
 
     private static final int MAX_SUBTYPE_LENGTH = 100; // Constrained by the database