+++ replacer-backend/src/test/java/es/bvalero/replacer/finder/replacement/finders/AcuteOFinderTest.java
@@ -38,7 +38,7 @@ class AcuteOFinderTest {
         assertEquals(1, replacements.size());
 
         Replacement rep = replacements.get(0);
-        assertEquals(ReplacementKind.COMPOSED, rep.getType().getKind());
+        assertEquals(ReplacementKind.STYLE, rep.getType().getKind());
         assertEquals(subtype, rep.getType().getSubtype());
         assertEquals(AcuteOFinder.ACUTE_O, rep.getText());
         assertEquals(AcuteOFinder.ACUTE_O, rep.getSuggestions().get(0).getText());
@@ -53,7 +53,7 @@ class AcuteOFinderTest {
         assertEquals(1, replacements.size());
 
         Replacement rep = replacements.get(0);
-        assertEquals(ReplacementKind.COMPOSED, rep.getType().getKind());
+        assertEquals(ReplacementKind.STYLE, rep.getType().getKind());
         assertEquals(SUBTYPE_ACUTE_O_WORDS, rep.getType().getSubtype());
         assertEquals(AcuteOFinder.ACUTE_O, rep.getText());
         assertEquals(7, rep.getStart());