+++ replacer-frontend/src/app/review/page/page-review.model.ts
@@ -6,7 +6,8 @@ export const kindLabel: { [key: number]: string } = {
   1: 'Personalizado',
   2: 'Ortografía',
   3: 'Compuestos',
-  4: 'Fechas'
+  4: 'Fechas',
+  5: 'Estilo'
 };
 
 export class ReviewOptions {